Revert "move `dev-cmd/bottle` methods to extend/os"
